Update project by following given feedback
