--- .eslintcache
@@ -1,1 +1,0 @@
-[{"/Users/apple/Desktop/Microverse/React/bookstore/src/components/BooksList.js":"1","/Users/apple/Desktop/Microverse/React/bookstore/src/App.js":"2","/Users/apple/Desktop/Microverse/React/bookstore/src/components/BooksForm.js":"3","/Users/apple/Desktop/Microverse/React/bookstore/src/actions/actionTypes.js":"4"},{"size":566,"mtime":1606301107157,"results":"5","hashOfConfig":"6"},{"size":363,"mtime":1606301193637,"results":"7","hashOfConfig":"6"},{"size":143,"mtime":1606298342338,"results":"8","hashOfConfig":"6"},{"size":93,"mtime":1606298749392,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"kexdeo",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/apple/Desktop/Microverse/React/bookstore/src/components/BooksList.js",[],"/Users/apple/Desktop/Microverse/React/bookstore/src/App.js",[],"/Users/apple/Desktop/Microverse/React/bookstore/src/components/BooksForm.js",[],"/Users/apple/Desktop/Microverse/React/bookstore/src/actions/actionTypes.js",[]]
